requirements.txt
fastapi
uvicorn
pyodbc
google-generativeai
python-dotenv